Retired Tiers for ease of use
Tiers have been replaced in the source files with the base cooldown which should make changes slightly easier to implement.
Also added sorting to the files for consistency and implemented a way to add new abilities to the database without directly editing the JSON files. (The current implementation doesn't ask you to double-check inputs, this will be implemented in the next commit.)