Schema fixes and updates
CooldownOverride, Scalar, and FlatIncrease properties of Override objects are no longer all required, and instead only one of them has to be present.

Added non-required ChargeBasedScaling property to Ability objects with additional documentation in the schema.

This commit also includes wording and clarity improvements to the schema as well as fixes to past mistakes I've made due to never having worked with schemas before.